pyproject.toml
[project]
name = "motus"
version = "0.1.0"
description = "Motus Agent Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "colorlog>=6.8.0",
    "fastapi>=0.128.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "openai>=2.30",
    "pydantic>=2.12.5",
    "rich>=13.0.0",
    "uvicorn>=0.40.0",
    "docker>=7.1.0",
    "kubernetes>=34.1.0",
    "mcp[cli]>=1.25.0",
    "paramiko>=4.0.0",
    "pip>=25.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "jsonref>=1.1.0",
    "tiktoken>=0.12.0",
    "backports-zstd>=1.3.0; python_version < '3.14'",
    "pathspec>=1.0.4",
    "tomli-w>=1.0.0",
]

[project.scripts]
motus = "motus.cli:main"

[project.optional-dependencies]
openai-agents = [
    "openai-agents>=0.0.10",
]
claude-agent-sdk = [
    "claude-agent-sdk>=0.1.0",
]
google-adk = [
    "google-adk>=1.27.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
motus = ["runtime/tracing/templates/*"]

[tool.uv.sources]
motus-deep-research = { workspace = true }

[tool.uv.workspace]
members = [
    ".",
    "services/*",
    "examples/deep_research",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "motus-deep-research",
    "pip>=25.3",
    "pre-commit>=4.5.1",
    "pytest-vcr>=1.0.2",
    "vcrpy>=8.1.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Fast unit tests (no external dependencies)",
    "integration: Integration tests using cassettes",
    "slow: Slow tests (real API calls or heavy operations)",
]

[tool.ruff.lint]
# E: pycodestyle errors, F: Pyflakes, I: Isort, W: pycodestyle warnings
select = ["E", "F", "I", "W"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["motus"]

[tool.ruff.lint.per-file-ignores]
"**/*.py" = ["E501", "E402"]
